Version 0.1 fixed.
The GUI layout was fixed to place all input parameters and graphs in their desired location.

Two additional windows were created to allow for the display of stats which occupied additional place on the main window squeezing the graphs. The other window was created to allow for better visualisation using the angualar graph because on the main window it is too small.

: